cmd: remove `{add,remove}-package` and `upgrade` commands
